+++ packages/core/src/index.ts
@@ -64,6 +64,7 @@ export * from './ide/ide-installer.js';
 export { getIdeInfo, DetectedIde } from './ide/detect-ide.js';
 export { type IdeInfo } from './ide/detect-ide.js';
 export * from './ide/constants.js';
+export * from './ide/types.js';
 
 // Export Shell Execution Service
 export * from './services/shellExecutionService.js';
